Rework scanner
The scanner now keeps in memory only the characters necessary
to interpret the next token (in the worst case this is still
the whole file, but in most cases this should not be the case).

Furthermore, the scanner now panics on invalid UTF-8 encodings,
preventing the consumer from having necessarily to deal with
those.